make file

+++ roboshop/components/common.sh
@@ -28,44 +28,41 @@ NodeJS(){
   yum install nodejs make gcc-c++ -y &>>$LOG
   stat $?
 
-
-  print "Add Roboshop "
-  id roboshop &>>$LOG
-  if [ $? -eq 0 ]; then
-    echo user roboshop already exist &>>$LOG
-  else
-    useradd roboshop &>>$LOG
-  fi
-  stat $?
-
-
   print "Download $COMPONENT_NAME"
   curl -s -L -o /tmp/${COMPONENT}.zip "https://github.com/roboshop-devops-project/${COMPONENT}/archive/main.zip" &>>$LOG
   stat $?
 
-  print "Remove Old Content"
-  rm -rf /home/roboshop/${COMPONENT}
-  stat $?
-
-  print "Extract $COMPONENT_NAME"
-  unzip -o -d /home/roboshop /tmp/${COMPONENT}.zip &>>$LOG
-  stat $?
-
-  print "Copy Content"
-  mv /home/roboshop/${COMPONENT}-main /home/roboshop/${COMPONENT}
-  stat $?
-
-  print "Install NodeJS Dependencies"
-  cd /home/roboshop/${COMPONENT}
-  npm install --unsafe-perm &>>$LOG
-  stat $?
-
+   print "Extract $COMPONENT_NAME"
+   unzip -o -d /home/roboshop /tmp/${COMPONENT}.zip &>>$LOG
+   stat $?
+   if [ "$1" == "/home/roboshop" ]; then
+       print "Remove Old Content"
+        rm -rf /home/roboshop/${COMPONENT}
+        stat $?
+        print "Copy Content"
+          mv /home/roboshop/${COMPONENT}-main /home/roboshop/${COMPONENT}
+          stat $?
+    fi
+  }
+ROBOSHOP_USER(){
+   print "Add Roboshop "
+    id roboshop &>>$LOG
+    if [ $? -eq 0 ]; then
+      echo user roboshop already exist &>>$LOG
+    else
+      useradd roboshop &>>$LOG
+    fi
+    stat $?
+SYSTEMD(){
   print "Fix App Permissions"
-  chown -R roboshop:roboshop /home/roboshop
-  stat $?
+    chown -R roboshop:roboshop /home/roboshop
+    stat $?
+}
+
   print "Update DNS records in SystemD config"
   sed -i -e 's/MONGO_DNSNAME/mongodb.roboshop.internal/' -e 's/REDIS_ENDPOINT/reddis.roboshop.internal/' -e 's/MONGO_ENDPOINT/mongodb.roboshop.internal/' -e 's/CATALOGUE_ENDPOINT/catalogue.roboshop.internal/' /home/roboshop/${COMPONENT}/systemd.service &>>$LOG
   stat $?
+
   print "Copy SystemD file"
   mv /home/roboshop/${COMPONENT}/systemd.service /etc/systemd/system/${COMPONENT}.service
   stat $?
@@ -73,7 +70,52 @@ NodeJS(){
   systemctl daemon-reload &>>$LOG && systemctl restart ${COMPONENT} &>>$LOG && systemctl enable ${COMPONENT} &>>$LOG
   stat $?
 
+PYTHON(){
+  print "Install Pyhton 3"
+  yum install python 36 gcc python3-devel -y &>>$LOG
+  stat $?
 
+  ROBOSHOP_USER
+  DOWNLOAD "/home/roboshop"
+
+  print "Install the dependencies"
+  cd /home/roboshop/${COMPONENT}
+  pip3 install -r requirements.txt &>>$LOG
+  stat $?
+
+  SYSTEMD
+}
+  MAVEN() {
+    print "Install Maven"
+    yum install maven -y &>>$LOG
+    stat $
+  }
+
+  ROBOSHOP_USER
+  DOWNLOAD "/home/roboshop"
+
+  print "Make Maven Package"
+  cd /home/roboshop/${COMPONENT}
+  mvn clean package &>>$LOG && mv target/shipping-1.0.jar shipping.jar &>>$LOG
+  stat $?
+
+  SYSTEMD
+}
+NodeJS() {
+  print "Install nodejs"
+  yum install nodejs make gcc-c++ -y &>>$LOG
+  stat $?
+
+  ROBOSHOP_USER
+
+  DOWNLOAD "/home/roboshop"
+
+  print "Install NodeJS Dependencies"
+  cd /home/roboshop/${COMPONENT}
+  npm install --unsafe-perm &>>$LOG
+  stat $?
+
+  SYSTEMD
 
 }
 
@@ -81,8 +123,7 @@ CHECK_MONGO_FROM_APP(){
   print "checking DB connections from app"
   sleep 5
   STAT=$(curl -s localhost:8080/health |  jq .mongo)
-  echo status = $STAT
-  if [ $STAT == "true" ]; then
+  if [ "$STAT" == "true" ]; then
     stat 0
   else
     stat 1
@@ -93,7 +134,7 @@ CHECK_REDDIS_FROM_APP(){
   sleep 5
   STAT=$(curl -s localhost:8080/health |  jq .redis)
   echo status = $STAT
-  if [ $STAT == "true" ]; then
+  if [ "$STAT" == "true" ]; then
     stat 0
   else
     stat 1
